requirements.txt
python-binance
python-dotenv
colorama
python-telegram-bot